+++ CHANGELOG.md
@@ -7,7 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
-### n/a
+### Added
+- Filename attribute for RpmUnit class
+- nsvca property for ModulemdUnit class
 
 ## [2.9.0] - 2021-06-15
 ### Added